[formerly 4e50d60552d4079becee1607b00783db633b92a2] [formerly 7892d4ba840bb86283d10adcdbda77005489a16f [formerly 8dea652e48ef897666dc4e25b6589a2bc2ad80d2] [formerly 7892d4ba840bb86283d10adcdbda77005489a16f [formerly 8dea652e48ef897666dc4e25b6589a2bc2ad80d2] [formerly 7892d4ba840bb86283d10adcdbda77005489a16f [formerly 8dea652e48ef897666dc4e25b6589a2bc2ad80d2] [formerly 74cae7a224253fd7056e9406922d7e876c85b157 [formerly 1d6f9cb8f7b8992852f16010bd5f77bcd040f025]]]] [formerly 0bbda52dcfacefe91cfb568698e0ef860da7ffe8 [formerly a457afe22a6f3b155eb62e9a16fbb7355c3833ba] [formerly 0bbda52dcfacefe91cfb568698e0ef860da7ffe8 [formerly a457afe22a6f3b155eb62e9a16fbb7355c3833ba] [formerly 0bbda52dcfacefe91cfb568698e0ef860da7ffe8 [formerly a457afe22a6f3b155eb62e9a16fbb7355c3833ba] [formerly 96fc8779411ccb827bd5efcefe494fedd35626ac [formerly 67446be728974547e39b838ec9f7f6e8d7448489]]]]]]




Former-commit-id: b0999d2159287d7d55ac96ec5f0a3b8a4fea0c29 [formerly aa13250f985f290b035f93797bf3e69ff750df3d]
Former-commit-id: 5c72b86ad2d514db7f552809fbf150f3ff07ba13 [formerly 189300b4f0ca2e3f5b5e7a283decd84cd948592a] [formerly 2e2d738467067e19ac6db642fecc4a4f80807ab7 [formerly 23c506b1d5af8285b43eab0bdcce48de1fe0ee66]]
Former-commit-id: 12eff12f673640478bce9e0416253bf63b7907af [formerly 554697a15788962b0c6a9027d871a3440d03d4a5]
Former-commit-id: e2d90fff55c15f643b74ae805f3bd8ff1fc40959

+++ plugins/inputs/trace/trace_skywalk_http.go
@@ -17,7 +17,7 @@ type SkyWalkLogData struct {
 }
 
 type SkyWalkLog struct {
-	Time int64
+	Time float64
 	Data []*SkyWalkLogData      `json:"data,omitempty"`
 }
